Refactor FileAnalysisCache
 - Split the generic cache implementation from the file-backing persistence
    strategy. This could probably work better as a composition rather than
    inheritance, but we will take it one step at a time.
 - Took the chance to improve error messages produces by FileAnalysisCache
