+++ pmd-core/src/test/java/net/sourceforge/pmd/cache/FileAnalysisCacheTest.java
@@ -29,6 +29,7 @@ public class FileAnalysisCacheTest {
     public TemporaryFolder tempFolder = new TemporaryFolder();
     
     private File unexistingCacheFile;
+    private File newCacheFile;
     private File emptyCacheFile;
     
     private File sourceFile;
@@ -36,6 +37,7 @@ public class FileAnalysisCacheTest {
     @Before
     public void setUp() throws IOException {
         unexistingCacheFile = new File(tempFolder.getRoot(), "non-existing-file.cache");
+        newCacheFile = new File(tempFolder.getRoot(), "pmd-analysis.cache");
         emptyCacheFile = tempFolder.newFile();
         sourceFile = tempFolder.newFile("Source.java");
     }
@@ -71,23 +73,28 @@ public class FileAnalysisCacheTest {
         cache.persist();
         assertTrue("Cache file doesn't exist after store", unexistingCacheFile.exists());
     }
-    
+
     @Test
     public void testStoreOnUnwritableFile() {
         emptyCacheFile.setWritable(false);
         final FileAnalysisCache cache = FileAnalysisCache.fromFile(emptyCacheFile);
         cache.persist();
+        // TODO : make proper assertions!
         assertTrue("Cache file doesn't exist after store", emptyCacheFile.exists());
     }
-    
+
     @Test
     public void testStoreSkipsFilesWithViolations() {
-        final FileAnalysisCache cache = FileAnalysisCache.fromFile(emptyCacheFile);
+        final FileAnalysisCache cache = FileAnalysisCache.fromFile(newCacheFile);
         cache.isUpToDate(sourceFile);
-        cache.ruleViolationAdded(new DummyRuleViolation(sourceFile));
+
+        final RuleViolation rv = mock(RuleViolation.class);
+        when(rv.getFilename()).thenReturn(sourceFile.getPath());
+
+        cache.ruleViolationAdded(rv);
         cache.persist();
-        
-        final FileAnalysisCache reloadedCache = FileAnalysisCache.fromFile(emptyCacheFile);
+
+        final FileAnalysisCache reloadedCache = FileAnalysisCache.fromFile(newCacheFile);
         assertFalse("Cache believes unmodified file with violations is up to date",
                 reloadedCache.isUpToDate(sourceFile));
     }
@@ -96,23 +103,23 @@ public class FileAnalysisCacheTest {
     public void testCacheValidityWithNoChanges() {
         final RuleSets rs = mock(RuleSets.class);
         final URLClassLoader cl = mock(URLClassLoader.class);
-        
-        setupCacheWithFiles(emptyCacheFile, rs, cl, sourceFile);
-        
-        final FileAnalysisCache reloadedCache = FileAnalysisCache.fromFile(emptyCacheFile);
+
+        setupCacheWithFiles(newCacheFile, rs, cl, sourceFile);
+
+        final FileAnalysisCache reloadedCache = FileAnalysisCache.fromFile(newCacheFile);
         reloadedCache.checkValidity(rs, cl);
         assertTrue("Cache believes unmodified file is not up to date without ruleset / classpath changes",
                 reloadedCache.isUpToDate(sourceFile));
     }
-    
+
     @Test
     public void testRulesetChangeInvalidatesCache() {
         final RuleSets rs = mock(RuleSets.class);
         final URLClassLoader cl = mock(URLClassLoader.class);
         
-        setupCacheWithFiles(emptyCacheFile, rs, cl, sourceFile);
+        setupCacheWithFiles(newCacheFile, rs, cl, sourceFile);
         
-        final FileAnalysisCache reloadedCache = FileAnalysisCache.fromFile(emptyCacheFile);
+        final FileAnalysisCache reloadedCache = FileAnalysisCache.fromFile(newCacheFile);
         when(rs.getChecksum()).thenReturn(1L);
         reloadedCache.checkValidity(rs, cl);
         assertFalse("Cache believes unmodified file is up to date after ruleset changed",
@@ -124,9 +131,9 @@ public class FileAnalysisCacheTest {
         final RuleSets rs = mock(RuleSets.class);
         final URLClassLoader cl = mock(URLClassLoader.class);
         
-        setupCacheWithFiles(emptyCacheFile, rs, cl, sourceFile);
+        setupCacheWithFiles(newCacheFile, rs, cl, sourceFile);
         
-        final FileAnalysisCache reloadedCache = FileAnalysisCache.fromFile(emptyCacheFile);
+        final FileAnalysisCache reloadedCache = FileAnalysisCache.fromFile(newCacheFile);
         when(cl.getURLs()).thenReturn(new URL[] { tempFolder.newFile().toURI().toURL(), });
         reloadedCache.checkValidity(rs, cl);
         assertTrue("Cache believes unmodified file is not up to date after classpath changed when no rule cares",
@@ -138,9 +145,9 @@ public class FileAnalysisCacheTest {
         final RuleSets rs = mock(RuleSets.class);
         final URLClassLoader cl = mock(URLClassLoader.class);
         
-        setupCacheWithFiles(emptyCacheFile, rs, cl, sourceFile);
+        setupCacheWithFiles(newCacheFile, rs, cl, sourceFile);
         
-        final FileAnalysisCache reloadedCache = FileAnalysisCache.fromFile(emptyCacheFile);
+        final FileAnalysisCache reloadedCache = FileAnalysisCache.fromFile(newCacheFile);
         when(cl.getURLs()).thenReturn(new URL[] { tempFolder.newFile().toURI().toURL(), });
         final net.sourceforge.pmd.Rule r = mock(net.sourceforge.pmd.Rule.class);
         when(r.usesDFA()).thenReturn(true);
@@ -152,28 +159,28 @@ public class FileAnalysisCacheTest {
 
     @Test
     public void testUnknownFileIsNotUpToDate() throws IOException {
-        final FileAnalysisCache cache = FileAnalysisCache.fromFile(unexistingCacheFile);
+        final FileAnalysisCache cache = FileAnalysisCache.fromFile(newCacheFile);
         assertFalse("Cache believes an unknown file is up to date",
                 cache.isUpToDate(sourceFile));
     }
 
     @Test
     public void testFileIsUpToDate() throws IOException {
-        setupCacheWithFiles(emptyCacheFile, mock(RuleSets.class), mock(ClassLoader.class), sourceFile);
+        setupCacheWithFiles(newCacheFile, mock(RuleSets.class), mock(ClassLoader.class), sourceFile);
         
-        final FileAnalysisCache cache = FileAnalysisCache.fromFile(emptyCacheFile);
+        final FileAnalysisCache cache = FileAnalysisCache.fromFile(newCacheFile);
         assertTrue("Cache believes a known, unchanged file is not up to date",
                 cache.isUpToDate(sourceFile));
     }
     
     @Test
     public void testFileIsNotUpToDateWhenEdited() throws IOException {
-        setupCacheWithFiles(emptyCacheFile, mock(RuleSets.class), mock(ClassLoader.class), sourceFile);
+        setupCacheWithFiles(newCacheFile, mock(RuleSets.class), mock(ClassLoader.class), sourceFile);
         
         // Edit the file
         Files.write(Paths.get(sourceFile.getAbsolutePath()), "some text".getBytes());
         
-        final FileAnalysisCache cache = FileAnalysisCache.fromFile(emptyCacheFile);
+        final FileAnalysisCache cache = FileAnalysisCache.fromFile(newCacheFile);
         assertFalse("Cache believes a known, changed file is up to date",
                 cache.isUpToDate(sourceFile));
     }
@@ -188,74 +195,5 @@ public class FileAnalysisCacheTest {
             cache.isUpToDate(f);
         }
         cache.persist();
-    }
-    
-    private static class DummyRuleViolation implements RuleViolation {
-        
-        private final File f;
-        
-        public DummyRuleViolation(final File f) {
-            this.f = f;
-        }
-
-        @Override
-        public boolean isSuppressed() {
-            return false;
-        }
-        
-        @Override
-        public String getVariableName() {
-            return null;
-        }
-        
-        @Override
-        public net.sourceforge.pmd.Rule getRule() {
-            return null;
-        }
-        
-        @Override
-        public String getPackageName() {
-            return null;
-        }
-        
-        @Override
-        public String getMethodName() {
-            return null;
-        }
-        
-        @Override
-        public String getFilename() {
-            return f.getPath();
-        }
-        
-        @Override
-        public int getEndLine() {
-            return 0;
-        }
-        
-        @Override
-        public int getEndColumn() {
-            return 0;
-        }
-        
-        @Override
-        public String getDescription() {
-            return null;
-        }
-        
-        @Override
-        public String getClassName() {
-            return null;
-        }
-        
-        @Override
-        public int getBeginLine() {
-            return 0;
-        }
-        
-        @Override
-        public int getBeginColumn() {
-            return 0;
-        }
     }
 }
